+++ com.unity.netcode.gameobjects/Runtime/Core/NetworkObject.cs
@@ -1741,6 +1741,8 @@ namespace Unity.Netcode
 
         private void OnDestroy()
         {
+            SceneManager.activeSceneChanged -= CurrentlyActiveSceneChanged;
+
             // Apply the is destroying flag
             SetIsDestroying();
 
@@ -2520,6 +2522,10 @@ namespace Unity.Netcode
         // If you couldn't find your parent, we put you into OrphanChildren set and every time we spawn another NetworkObject locally due to replication,
         // we call CheckOrphanChildren() method and quickly iterate over OrphanChildren set and see if we can reparent/adopt one.
         internal static HashSet<NetworkObject> OrphanChildren = new HashSet<NetworkObject>();
+#if UNITY_EDITOR
+        [RuntimeInitializeOnLoadMethod(RuntimeInitializeLoadType.SubsystemRegistration)]
+        private static void ResetStaticsOnLoad() => OrphanChildren = new HashSet<NetworkObject>();
+#endif
 
         internal bool ApplyNetworkParenting(bool removeParent = false, bool ignoreNotSpawned = false, bool orphanedChildPass = false, bool enableNotification = true)
         {