+++ com.unity.netcode.gameobjects/Runtime/Messaging/MessageDelivery.cs
@@ -1,7 +1,6 @@
 using System;
 using System.Collections.Generic;
 using Unity.Netcode;
-using UnityEditor;
 using UnityEngine;
 
 internal static class MessageDelivery
@@ -15,17 +14,21 @@ internal static class MessageDelivery
     /// when sending the message via public API.
     /// - Skip the time sync messages since it has always used unreliable network delivery.
     /// </summary>
-    private static HashSet<NetworkMessageTypes> s_SkipMessageTypes = new HashSet<NetworkMessageTypes>(){
+    private static readonly HashSet<NetworkMessageTypes> k_SkipMessageTypes = new HashSet<NetworkMessageTypes>(){
         NetworkMessageTypes.NamedMessage, NetworkMessageTypes.Unnamed};
 
-    [RuntimeInitializeOnLoadMethod]
+    [RuntimeInitializeOnLoadMethod(RuntimeInitializeLoadType.SubsystemRegistration)]
     private static void OnApplicationStart()
     {
+#if UNITY_EDITOR
+        s_MessageToDelivery = new Dictionary<NetworkMessageTypes, NetworkDelivery>();
+        s_MessageToMessageType = new Dictionary<Type, NetworkMessageTypes>();
+#endif
         UpdateMessageTypes();
     }
 
     /// <summary>
-    /// FIrst pass at providing an easier path to configuring the network
+    /// First pass at providing an easier path to configuring the network
     /// delivery type for the message type.
     /// TODO: Once <see cref="NetworkMessageManager"/> coalesces all reliable messages
     /// and/or organizes by a more unified order of operation tracking built into the
@@ -40,7 +43,7 @@ internal static class MessageDelivery
         foreach (var messageTypeObject in networkMessageTypes)
         {
             var messageType = (NetworkMessageTypes)messageTypeObject;
-            if (s_SkipMessageTypes.Contains(messageType))
+            if (k_SkipMessageTypes.Contains(messageType))
             {
                 continue;
             }
@@ -56,9 +59,13 @@ internal static class MessageDelivery
         MessageDeliveryType<ConnectionApprovedMessage>.Initialize();
         MessageDeliveryType<CreateObjectMessage>.Initialize();
         MessageDeliveryType<DestroyObjectMessage>.Initialize();
+        MessageDeliveryType<DisconnectReasonMessage>.Initialize();
+        MessageDeliveryType<NamedMessage>.Initialize();
+        MessageDeliveryType<UnnamedMessage>.Initialize();
         MessageDeliveryType<NetworkTransformMessage>.Initialize();
         MessageDeliveryType<NetworkVariableDeltaMessage>.Initialize();
         MessageDeliveryType<ParentSyncMessage>.Initialize();
+        MessageDeliveryType<ProxyMessage>.Initialize();
         // RpcMessage.cs
         {
             MessageDeliveryType<RpcMessage>.Initialize();
@@ -71,18 +78,10 @@ internal static class MessageDelivery
         MessageDeliveryType<TimeSyncMessage>.Initialize();
     }
 
-#if UNITY_EDITOR
-    [InitializeOnLoadMethod]
-    [InitializeOnEnterPlayMode]
-    private static void OnEnterPlayMode()
-    {
-        UpdateMessageTypes();
-    }
-#endif
     internal static NetworkDelivery GetDelivery(Type type)
     {
         // Return the default if not registered or null
-        if (type == null || s_SkipMessageTypes.Contains(s_MessageToMessageType[type]))
+        if (type == null || k_SkipMessageTypes.Contains(s_MessageToMessageType[type]))
         {
             return NetworkDelivery.ReliableFragmentedSequenced;
         }
@@ -91,7 +90,7 @@ internal static class MessageDelivery
 
     internal static NetworkDelivery GetDelivery(NetworkMessageTypes messageType)
     {
-        if (s_SkipMessageTypes.Contains(messageType))
+        if (k_SkipMessageTypes.Contains(messageType))
         {
             throw new Exception($"{messageType} is not registered in the message type to network delivery map!");
         }